Rename example-checkout to example-basic
