+++ example-basic/build.gradle
@@ -11,7 +11,7 @@ android {
     buildToolsVersion rootProject.buildToolsVersion
 
     defaultConfig {
-        applicationId "net.optile.example.checkout"
+        applicationId "net.optile.example.basic"
         minSdkVersion rootProject.minSdkVersion
         targetSdkVersion rootProject.targetSdkVersion
         versionCode rootProject.versionCode
