+++ example-basic/proguard-rules.pro
