+++ example-basic/src/androidTest/java/net/optile/payment/UITestSuite.java
