+++ example-basic/src/androidTest/res/raw/preset.json
