+++ example-basic/src/main/res/drawable/custom_logobackground.xml
