+++ example-basic/src/main/res/drawable/response_background.xml
