+++ example-basic/src/main/res/font/lato_regular.ttf
