+++ example-basic/src/main/res/layout/activity_basic.xml
