+++ example-basic/src/main/res/raw/customgroups.json
