+++ example-basic/src/main/res/raw/customvalidations.json
