+++ example-basic/src/main/res/values/colors.xml
