+++ example-basic/src/main/res/values/dimens.xml
