+++ example-basic/src/main/res/values/styles.xml
