+++ example-basic/src/main/res/values/themes.xml
